Use Middleware to inject object

+++ app/src/Routes/RolesRoutes.php
@@ -27,6 +27,7 @@ use UserFrosting\Sprinkle\Admin\Controller\Role\RolePageAction;
 use UserFrosting\Sprinkle\Admin\Controller\Role\RolePermissionsSprunje;
 use UserFrosting\Sprinkle\Admin\Controller\Role\RolesPageAction;
 use UserFrosting\Sprinkle\Admin\Controller\Role\RoleUsersSprunje;
+use UserFrosting\Sprinkle\Admin\Middlewares\RoleInjector;
 
 /*
  * Routes for administrative role management.
@@ -37,25 +38,25 @@ class RolesRoutes implements RouteDefinitionInterface
     {
         $app->group('/roles', function (RouteCollectorProxy $group) {
             $group->get('', RolesPageAction::class)->setName('uri_roles');
-            $group->get('/r/{slug}', RolePageAction::class);
+            $group->get('/r/{slug}', RolePageAction::class)->add(RoleInjector::class);
         })->add(AuthGuard::class); //->add(new NoCache());
 
         $app->group('/api/roles', function (RouteCollectorProxy $group) {
-            $group->delete('/r/{slug}', RoleDeleteAction::class);
+            $group->delete('/r/{slug}', RoleDeleteAction::class)->add(RoleInjector::class);
             $group->get('', [RolesPageAction::class, 'sprunje']);
             // $group->get('/r/{slug}', 'UserFrosting\Sprinkle\Admin\Controller\RoleController:getInfo');
-            $group->get('/r/{slug}/permissions', RolePermissionsSprunje::class);
-            $group->get('/r/{slug}/users', RoleUsersSprunje::class);
+            $group->get('/r/{slug}/permissions', RolePermissionsSprunje::class)->add(RoleInjector::class);
+            $group->get('/r/{slug}/users', RoleUsersSprunje::class)->add(RoleInjector::class);
             $group->post('', RoleCreateAction::class);
-            $group->put('/r/{slug}', RoleEditAction::class);
+            $group->put('/r/{slug}', RoleEditAction::class)->add(RoleInjector::class);
             // $group->put('/r/{slug}/{field}', 'UserFrosting\Sprinkle\Admin\Controller\RoleController:updateField');
         })->add(AuthGuard::class); //->add(new NoCache());
 
         $app->group('/modals/roles', function (RouteCollectorProxy $group) {
-            $group->get('/confirm-delete', RoleDeleteModal::class)->setName('modal.roles.delete');
+            $group->get('/confirm-delete', RoleDeleteModal::class)->add(RoleInjector::class)->setName('modal.roles.delete');
             $group->get('/create', RoleCreateModal::class)->setName('modal.roles.create');
-            $group->get('/edit', RoleEditModal::class)->setName('modal.roles.edit');
-            $group->get('/permissions', RoleEditPermissionsModal::class)->setName('modal.roles.permissions');
+            $group->get('/edit', RoleEditModal::class)->add(RoleInjector::class)->setName('modal.roles.edit');
+            $group->get('/permissions', RoleEditPermissionsModal::class)->add(RoleInjector::class)->setName('modal.roles.permissions');
         })->add(AuthGuard::class); //->add(new NoCache());
     }
 }
